# Conflicts:
#	ICT_MARKET/src/main/webapp/users/login_form.jsp

+++ ICT_MARKET/src/main/java/service/UserInfoDeleteService.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import kr.co.ict.UserDAO;
 
@@ -12,14 +13,18 @@ public class UserInfoDeleteService implements boardInterface_Service{
 
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+		String user_id = request.getParameter("user_id");
 		String struNum = request.getParameter("user_num");
 		int uNum = Integer.parseInt(struNum);
-		
-		// 1. dao»ý¼º
+		HttpSession session = request.getSession();
+		String sId = (String)session.getAttribute("session_id");
+		// 1. daoï¿½ï¿½ï¿½ï¿½
+		if(sId == user_id) {
 		UserDAO dao = UserDAO.getInstance();
-		// 2. »èÁ¦·ÎÁ÷ ½ÇÇà
+		// 2. ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ï¿½ ï¿½ï¿½ï¿½ï¿½
 		dao.deleteUser(uNum);
-		
+		session.invalidate();
+		}
 	}
 
 }
